requirements.txt
flask>=3.0
cryptography>=42.0
click>=8.1
pyyaml>=6.0
colorama>=0.4
tabulate>=0.9
jinja2>=3.1
bcrypt>=4.0
python-dotenv>=1.0
flask-limiter>=3.5
